requirement.txt
absl-py==0.6.1
astor==0.7.1
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
certifi==2018.10.15
cffi==1.11.5
cycler==0.10.0
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==2.2.3
mkl-fft==1.0.6
mkl-random==1.0.2
mock==2.0.0
numpy==1.15.4
olefile==0.46
opencv-contrib-python==3.4.3.18
pbr==5.1.1
Pillow==5.3.0
protobuf==3.6.1
pycparser==2.19
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
scipy==1.1.0
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.5.3
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
torch==0.4.1
torchvision==0.2.1
tornado==5.1.1
Werkzeug==0.14.1
